Added documentation to all TODOList module components.
Created a new CategoryToggle component for filtering the TODO list view via categories.
