+++ src/components/TODOList/ListView.js
@@ -1,7 +1,12 @@
 import React from 'react';
 import TodoItem from './TODOItem';
 
+// The ListView component is responsible for rendering the list of visible TODOitems.
 const ListView = ({ visibleTODOs }) => {
+
+    // Since the props visibleTODOs is not mutable, a new array sortedTodos is created using
+    // the spread operator [...visibleTODOs]. The sort method is then called on sortedTodos
+    // to sort the TODOItems based on their dueDate property in ascending order.
     let sortedTodos = [...visibleTODOs].sort((a, b) => a.dueDate - b.dueDate);
 
     return (
