+++ src/components/TODOList/TODOListMainView.js
@@ -3,12 +3,18 @@ import TODOListViewer from './TODOListViewer';
 import ControlPanel from './ControlPanel';
 import {useSelector} from "react-redux";
 
+// This TODOListMainView component represents the main view of the TODOlist,
+// displaying the list of TODOItems based on the selected category.
 const TODOListMainView = () => {
-    // The destructuring assignment is used to directly extract the
-    // TODOList state from the todoReducer slice
+    // The useSelector hook extracts the TODOList array from the todoListSlice
+    // slice in the rootReducer.
     const { TODOList } = useSelector((state) => state.todoReducer);
+
+    // The selectedCategory state variable keeps track of the currently selected
+    // category in the TODOList.
     const [selectedCategory, setSelectedCategory] = useState('');
 
+    // The getVisibleTODOs function is defined to filter the TODOList based on the selected category.
     const getVisibleTODOs = () => {
         if (!selectedCategory) {
             return TODOList;
