+++ src/store/todoListSlice.js
@@ -1,5 +1,8 @@
 import {createSlice, nanoid} from "@reduxjs/toolkit";
 
+// TODOListSlice sets up the initial redux state, defines reducers for adding,
+// editing, and deleting TODO items
+
 const INITIAL_STATE = {
     categories: [
         'biology',
@@ -71,12 +74,15 @@ const TODOListSlice = createSlice({
     name: "todoList",
     initialState: INITIAL_STATE,
     reducers: {
+        // The addTODO reducer is responsible for adding a new TODO item to the
+        // TODOList array in the state.
         addTODO: {
             reducer: (state, action) => {
                 state.TODOList.push(action.payload);
             },
-            // The addTODO uses the prepare function to generate and initialize the payload with
-            // the necessary fields PRIOR to applying the reducer to it.
+            // The prepare function is a pre-reducer that prepares the payload
+            // for the addTODO reducer by generating a unique ID using
+            // nanoid() and merging it with the provided todo object.
             prepare: (todo) => {
                 return {
                     payload: {
@@ -116,7 +122,7 @@ const TODOListSlice = createSlice({
         addCategory: (state, action) => {
             const categoryIndex = state.categories.findIndex((item) => item === action.payload);
 
-            // Only add payload to state.categories only if category does not already
+            // Only add payload to state.categories if category does not already
             // exist in the array
             if (categoryIndex === -1) {
                 state.categories.push(action.payload);
@@ -126,6 +132,8 @@ const TODOListSlice = createSlice({
     },
 });
 
+// Exports the action creators generated by the createSlice function as named exports.
+// These action creators can be used to dispatch actions to modify the TODO list state.
 export const {
     addTODO,
     editTODO,
@@ -134,5 +142,7 @@ export const {
     addCategory,
 } = TODOListSlice.actions;
 
+// The TODOListSlice.reducer is exported as the default export, representing the reducer
+// function for this slice.
 export default TODOListSlice.reducer;
